cc: let a designator list walk back out of a nested aggregate; libap: readlinkat returns ssize_t
Two unrelated failures in diffutils' diff.c.

  diff.c:1265 argument prototype mismatch
    "IND FUNC(INT, IND CONST CHAR, IND CHAR, UVLONG) INT" for
    "IND FUNC(INT, IND CONST CHAR, IND CHAR, UVLONG) VLONG"

POSIX.1-2008 gives readlinkat an ssize_t return, as readlink already has
in <sys/stat.h>. APE declared and defined it int, which gnulib caught
because careadlinkat takes the reader as a parameter and typechecks it.
It was also an ABI bug of the same family as the size_t one: on amd64
the callee wrote four bytes where an ssize_t caller reads eight.
readlinkat is the only *at function POSIX gives a non-int return, and
the only one in at_functions.c that was wrong.

  diff.c:1412 structure element not found desc

This is a compiler bug, in designator lists that leave a nested
aggregate and come back:

  struct comparison cmp = { .file[0].desc = ..., .file[1].desc = ...,
                            .file[0].stat.st_size = ..., .parent = ... };

The grammar builds these chains correctly and init1 descends to the
leftmost designator and trims as it matches, so entering file[0].desc
works. Coming back out did not. Having initialised file[0], init1 was
still inside struct file_data when it peeked at .file[1].desc, found no
member named file, and reported the outermost name in the chain -- desc.

A designator that names nothing here is an error only at a level with
its own braces. exflag distinguishes them: it is 0 for the outermost
aggregate and for every level reached through doinit, which is the brace
case, and 1 when init1 was entered implicitly. So when exflag is set,
return instead of diagnosing and let the enclosing level match it. A
designator that matches nothing anywhere still reaches the outermost
level, where exflag is 0, and is reported there as before.

The array case has to cooperate, or it hands the same designator to the
element type again. init1 consuming nothing is the signal; peekinit has
no side effects, so comparing it across the call is enough. Guarded on
the pending node being a designator, so ordinary element initialisers,
brace-nested ones and the a == Z case are untouched.

+++ sys/include/ape/unistd.h
@@ -184,7 +184,7 @@ int fstatat(int dirfd, const char *path, struct stat *buf, int flags);
 int unlinkat(int dirfd, const char *path, int flags);
 int mkdirat(int dirfd, const char *path, mode_t mode);
 int renameat(int olddirfd, const char *old, int newdirfd, const char *new);
-int readlinkat(int dirfd, const char *path, char *buf, size_t n);
+ssize_t readlinkat(int dirfd, const char *path, char *buf, size_t n);
 int symlinkat(const char *target, int dirfd, const char *linkpath);
 int faccessat(int dirfd, const char *path, int mode, int flags);
 int fchownat(int dirfd, const char *path, uid_t uid, gid_t gid, int flags);
